refac(AbstractCartHandler): Refatora App.js
Para ter um código mais limpo, as funcionalidades
destinadas a gerenciar o cart presentes em App.js
foram transferidas para a classe/componente abstrata
AbstractCartHandler. App agora herda suas propriedades
e métodos.
